% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{process_data}
\alias{process_data}
\title{Preprocess data frame}
\usage{
process_data(data = NULL, x = NULL, y = NULL, group = NULL)
}
\arguments{
\item{data}{A data frame}

\item{x}{string value of column name containing values on the x-axis}

\item{y}{string value of column name containing values on the y-axis}

\item{group}{string value of column name for grouping}
}
\description{
Preprocess data frame
}
