% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{process_data}
\alias{process_data}
\title{Preprocess data frame}
\usage{
process_data(data, x, y)
}
\arguments{
\item{data}{A data frame}

\item{x}{column name containing values on the x-axis}

\item{y}{column name containing values on the y-axis}
}
\description{
Preprocess data frame
}
